Implementa fluxo de retirada, devolução e cancelamento de locação — adiciona StatusInvalidoException e os métodos confirmarRetirada, confirmarDevolucao e cancelarLocacao no LocacaoService; renomeia status JA_RETIRADO para RETIRADO e altera quilometragem do Veiculo para Double

+++ src/main/java/io/github/arthursilvagbs/Locacao/de/Carros/entity/Veiculo.java
@@ -60,7 +60,7 @@ public class Veiculo {
 
     @Setter
     @Column(precision = 7, scale = 1, nullable = false)
-    private Integer quilometragem;
+    private Double quilometragem;
 
     @Setter
     @OneToMany(mappedBy = "veiculo")
